Implement filterable getters for RSVP counts | #61992

+++ src/Tribe/RSVP/Attendance_Totals.php
@@ -5,6 +5,10 @@
  *
  * Also has the capability to print this information as HTML, intended for
  * use in the attendee summary screen.
+ *
+ * Note that the various counts are determined during instantiation, therefore
+ * if the status of one or more RSVPs are modified *after* creating an object
+ * of this type
  */
 class Tribe__Tickets__RSVP__Attendance_Totals {
 	protected $event_id = 0;
@@ -85,6 +89,11 @@ class Tribe__Tickets__RSVP__Attendance_Totals {
 		";
 	}
 
+	/**
+	 * The total number of RSVPs received for this event.
+	 *
+	 * @return int
+	 */
 	public function get_total_rsvps() {
 		/**
 		 * Returns the total RSVP count for an event.
@@ -93,9 +102,15 @@ class Tribe__Tickets__RSVP__Attendance_Totals {
 		 * @param int $original_total_rsvps
 		 * @param int $event_id
 		 */
-		return apply_filters( 'tribe_tickets_rsvp_get_total_rsvps', $this->total_rsvps, $this->total_rsvps, $this->event_id );
+		return (int) apply_filters( 'tribe_tickets_rsvp_get_total_rsvps', $this->total_rsvps, $this->total_rsvps, $this->event_id );
 	}
 
+	/**
+	 * The total number of RSVPs for this event that indicate they are
+	 * going.
+	 *
+	 * @return int
+	 */
 	public function get_total_going() {
 		/**
 		 * Returns the total going count for an event.
@@ -104,9 +119,15 @@ class Tribe__Tickets__RSVP__Attendance_Totals {
 		 * @param int $original_total_going
 		 * @param int $event_id
 		 */
-		return apply_filters( 'tribe_tickets_rsvp_get_total_going', $this->total_going, $this->total_going, $this->event_id );
+		return (int) apply_filters( 'tribe_tickets_rsvp_get_total_going', $this->total_going, $this->total_going, $this->event_id );
 	}
 
+	/**
+	 * The total number of RSVPs for this event that indicate they are
+	 * not going.
+	 *
+	 * @return int
+	 */
 	public function get_total_not_going() {
 		/**
 		 * Returns the total not going count for an event.
@@ -115,6 +136,6 @@ class Tribe__Tickets__RSVP__Attendance_Totals {
 		 * @param int $original_total_not_going
 		 * @param int $event_id
 		 */
-		return apply_filters( 'tribe_tickets_rsvp_get_total_not_going', $this->total_not_going, $this->total_not_going, $this->event_id );
+		return (int) apply_filters( 'tribe_tickets_rsvp_get_total_not_going', $this->total_not_going, $this->total_not_going, $this->event_id );
 	}
 }
